+++ stories/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import icons from '../src/Icon/icons.json';
 import styled from 'styled-components';
-
+import { withKnobs, text, select } from '@storybook/addon-knobs';
+import { withInfo } from '@storybook/addon-info';
 import { storiesOf } from '@storybook/react';
-
 import Icon from '../src/Icon';
 
 const Container = styled.div`
@@ -19,8 +19,9 @@ const styles = {
 }
 
 const stories = storiesOf('Icons', module);
+stories.addDecorator(withKnobs);
 
-stories.add('all icons', () => {
+stories.add('all icons', withInfo()(() => {
         const iconsElements = icons.map( (icon, index) => {
           const Box = styled.div`
           border: 1px solid #eee;
@@ -29,6 +30,7 @@ stories.add('all icons', () => {
           display: flex;
           height: 100px;
           justify-content: center;
+          align-items: center;
           margin: 5px;
           padding: 20px 10px
           text-align: center;
@@ -41,19 +43,17 @@ stories.add('all icons', () => {
             font-size: 0.6rem;
             max-width: 90px;
             position: absolute;
-            margin-top: 70px;
-            opacity: 0;
+            margin-top: 2.5em;
+            opacity: 0.4;
           }
           &:hover:after {
             opacity: 1;
           }
         `
-          return <Box key={index} style={ styles.iconBox }>
+          return <Box key={index}>
             <Icon
-              style={styles.icon}
               description={icon.title}
               name={icon.title.split(' ').join('-')}
-              fill='black'
               size='medium'
             />
           </Box>
@@ -66,11 +66,19 @@ stories.add('all icons', () => {
           </div>
         );
       }
+  ));
+
+stories.add('Colored icon', () => {
+  return (
+    <h2 style={ { 'color': 'red' } }>
+      Test Me Red <Icon name='bell' size='large' />
+    </h2>
   );
+});
 
 stories.add('large icon', () => {
   return (
-    <Icon name='bell' size='large' />
+    <Icon name='bell' size='large' fill={ text('color', 'hotpink') } />
   );
 });